Cater to stock splits (post 2022)
Stock splits are not present in trading-212 exports at the moment but it
is a feature request. This is a stop gap and needs to be reviewed per
tax report generation.

More info on potentially adding stock splits for exports: https://community.trading212.com/t/missing-information-in-history-statements/48201
For now use: https://companiesmarketcap.com/eur/