Use sell record conversion when buy record is diff and has exchange rate of 1
better currency tests

add tests for cases where buy and sale were in separate currencies. We assume that sale was done in buy currency and then converted and reported in the csv. So the sale's exchange rate is used to calculate appropriate buys and base caculations off of this new number instead